// interfaces/LolRateDto.ts
export interface LolRateDto {
  id: number

  championName: string

  iconUrl: string

  role: string

  opggPick: number

  opggWin: number

  opggAvg: number

  opggUpdatedAt: string

  lolgraphsPick: number

  lolgraphsWin: number

  lolgraphsAvg: number

  lolgraphsUpdatedAt: string

  lolalyticsPick: number

  lolalyticsWin: number

  lolalyticsAvg: number

  lolalyticsUpdatedAt: string

  blitzPick: number

  blitzWin: number

  blitzAvg: number

  blitzUpdatedAt: string

  createdAt: string

  updatedAt: string
}